Move trait `BufferLike` to module `buffer` at crate root
