+++ pythonx/jieba_vim_rs_core/src/buffer.rs
@@ -1,0 +1,50 @@
+// Copyright 2024-2025 Kaiwen Wu. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy
+// of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
+/// Any type that resembles a Vim buffer.
+pub trait BufferLike {
+    type Error;
+
+    /// Get the line at line number `lnum` (1-indexed).
+    fn getline(&self, lnum: usize) -> Result<String, Self::Error>;
+
+    /// Get the total number of lines in the buffer.
+    fn lines(&self) -> Result<usize, Self::Error>;
+}
+
+#[cfg(test)]
+impl BufferLike for Vec<&'static str> {
+    type Error = ();
+
+    fn getline(&self, lnum: usize) -> Result<String, Self::Error> {
+        self.get(lnum - 1).map(|s| s.to_string()).ok_or(())
+    }
+
+    fn lines(&self) -> Result<usize, Self::Error> {
+        Ok(self.len())
+    }
+}
+
+#[cfg(test)]
+impl BufferLike for Vec<String> {
+    type Error = ();
+
+    fn getline(&self, lnum: usize) -> Result<String, Self::Error> {
+        self.get(lnum - 1).map(|s| s.to_string()).ok_or(())
+    }
+
+    fn lines(&self) -> Result<usize, Self::Error> {
+        Ok(self.len())
+    }
+}
