+++ pythonx/jieba_vim_rs_core/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2024 Kaiwen Wu. All Rights Reserved.
+// Copyright 2024-2025 Kaiwen Wu. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not
 // use this file except in compliance with the License. You may obtain a copy
@@ -12,6 +12,9 @@
 // License for the specific language governing permissions and limitations
 // under the License.
 
+mod buffer;
 pub mod motion;
 pub mod token;
 mod utils;
+
+pub use buffer::BufferLike;
